feat(decode): impl `TryFrom<&ResolvedFunction> for TraceFactory`
